Update test for 0.18
Elm 0.18 expect:
* The test folder to be named `test`
* The file to execute to be named Main.elm

Also ran `elm-format --upgrade` on the test, as some notation are not
valid anymore (and there was too many to manually change). This lead the
elm-check's test to be barelly readable anymore. (Since the backtick
notation is not valid anymore).

Note that elm-check is deprecated and should be replace by the `fuzz`
test from elm-test instead.

Since a lot changed in elm-test and elm-check, I used the
`rtfeldman/legacy-elm-test` package which can be use has a dropin
replacement to make your test pass until an upgrade to the last syntax
appen.

The test now run on nodejs instead of the browser, thus running elm-test
on the command line return the test suite status.